feat(linux): linux support for t-rec
 - make sure the generated images are visible by adjusting the alpha channel if it is missing (ubuntu: unity behaves this way)
 - cleanup unused imports
 - add env logger
 - grab a screenshot by window id
 - verifies the transparent border portion of the image and adjust the margins so that no transparent stuff is kept (tested on ubuntu)

+++ src/macos/screenshot.rs
@@ -5,13 +5,13 @@ use core_graphics::image::CGImageRef;
 use image::flat::SampleLayout;
 use image::{ColorType, FlatSamples};
 
-pub fn capture_window_screenshot(win_id: u32) -> Result<ImageOnHeap> {
+pub fn capture_window_screenshot(win_id: u64) -> Result<ImageOnHeap> {
     let (w, h, channels, raw_data) = {
         let image = unsafe {
             CGDisplay::screenshot(
                 CGRectNull,
                 kCGWindowListOptionIncludingWindow | kCGWindowListExcludeDesktopElements,
-                win_id,
+                win_id as u32,
                 kCGWindowImageNominalResolution
                     | kCGWindowImageBoundsIgnoreFraming
                     | kCGWindowImageShouldBeOpaque,
